move good search jsx into search component
