+++ client/src/views/Home.tsx
@@ -1,36 +1,19 @@
-import makeStyles from '@mui/styles/makeStyles';
-
 import Search from '../components/Search';
 import BannerSection from '../components/BannerSection';
 import NeedsCarousel from '../components/NeedsCarousel';
 import DonationsCarousel from '../components/DonationsCarousel';
 import FAQs from '../components/FAQs';
 
-const useStyles = makeStyles(() => ({
-  needsAndOffers: {
-    padding: '10%',
-  },
-  makeAPostButton: {
-    marginLeft: '15px',
-  },
-  makeAPostLink: {
-    textDecoration: 'none',
-  },
-  searchContainer: {
-    maxWidth: '980px',
-    width: '70%',
-    margin: '30px auto',
-  },
-}));
+// searchContainer: {
+//   maxWidth: '980px',
+//   width: '70%',
+//   margin: '30px auto',
+// },
 
 function Home(): JSX.Element {
-  const classes = useStyles();
-
   return (
     <>
-      <div className={classes.searchContainer}>
-        <Search />
-      </div>
+      <Search />
       <BannerSection />
       <NeedsCarousel label={'Recent Needs From Nonprofits'} />
       <DonationsCarousel label={'Recent Offers From Users'} />
